add git ignore changes.txt and fix

+++ functions.sh
@@ -44,7 +44,7 @@ successMessages(){
 
 # Función para comprobar si el archivo "changes.txt" existe y eliminar su contenido
 function deleteChangesFileContent() {
-    if [ -f "changes.txt" ]; then
+    if [ -f "changes.txt" ] && [ -s "changes.txt" ]; then
         # Elimina el contenido del archivo
         echo "" > changes.txt
     fi
